requirements.txt
python-dotenv
pyTelegramBotAPI
requests
requests-ntlm
schedule
ntlm-auth
numpy
pandas
Flask
openpyxl
bs4
lxml
telegraph
psycopg2-binary
psycopg2
psutil